Switch images using SCN 10 to use new location
When we first built SCN 10, we used an ED25519 GPG signing key for the
packages.  It turns out that this doesn't work in FIPS mode, so we have
switched to an RSA 4096 GPG signing key.  To ensure that updates continue
to work seamlessly for those who installed systems using the original key,
we have transitioned to a new location for packages signed with the new
key.

This commit updates the SCN image build process to use the new location.

This also fixes a minor repo naming issue (probably only relevant in the
logs).

Signed-off-by: Jonathan Dieter <jdieter@ciq.com>